Fall back to --selection in Wayland window mode
Temporary "fix" for #28 until either Slurp gets window selection
possibility, or we can add something in NextShot directly.

The issue with the latter is there seems to be no way to distinguish
between running NextShot in a terminal vs from a keybind, othweise we
could just output a list of windows and prompt the user to select one.