Fix bash error (again)
While 0.8.1 fixed the unary operator issue, it introduced a new one
where it was expecting an integer expression instead.

This reverts the quotes and instead checks if the $choice variable is
empty. If it is, it resets to -1 before the next iteration to ensure its
value will never be blank.