Abort on errors
Splits pipeline in half so that we can use set some options that, when
combined, will force the script to exit when there's an error in the
last command of a pipeline, even when said pipeline is in a function.

`-e` makes it exit, `-o pipefail` adds the magic for failing when last
command of pipe fails, and `-E` makes the set options be inherited by
functions etc.

Then adds an ERR trap to tell the user what happened, but this is still
pretty dumb at this point. Could use $BASH_COMMAND to echo the last
command ran, but that would be `filename=$(...)` rather than simply
`nc_upload` or pointing to the curl command itself.