Handle HTTP errors on upload
Tells curl to write the HTTP response code to stdout and saves it to the
$respCode variable before checking to ensure it's a 201 response. If
it's not, we echo a useful error message to stderr and exit.

Also redirects the usual output to /dev/null. When the request works as
expected there won't be any output, but if there *is* an error then we
don't want the standard HTML interpretation of it dumped to stdout or it
will interfere with the if statement.