Switch cli config to use abort helper
Moves yad/manual config message into abort_config and calls that instead
of exiting directly to benefit from the 'aborted by user' output.

Also inverts the "if answer is yes" check to abort earlier, and moves
the empty echo to be immediately after the read command so that there's
always a newline after the answer to the "create config?" question.