Rework config confirmation dialog to fix #27
Saves output of yad to a variable so we can abort_config when it fails,
fixing the issue where the ERR trap would catch it (and previously it
would simply not be caught at all).

Also moves the mkdir after yad to ensure config still happens on next
run if a user decides to hit escape in the confirmation window.