Immediately exit if missing vital config options
The `:` is merely an alias to `false`, ie it does nothing. This way we
can have our checks without actually having to reassign variables
that're already assigned.

In the parameter expansion, `:?` will `exit 1` if the var is null or not
set, and output the following string to the terminal with the error, eg:
`/usr/local/bin/nextshot: line 184: password: Required config option...`