Init cache with the rest of the setup stuff
Moves _CACHE_DIR to the top of the script with the rest of the globals
and makes them readonly, since they'll never be changed. This'll make
sure they can't be overridden from a user's config file.

That just leaves one line in init_cache so instead it can move to setup.