Handle unhandled and optless args
If an option is passed that's defined as an opt/longopt, but not checked
for in the parse loop, the script will run forever. It currently occurs
with -V because -V|--version hasn't been moved yet, but in reality it
should never happen at all.

Also added a case to deal with the `--` special option such that $1 gets
reset. Thanks to enhanced getopt, `--` will *always* be the last option
so we can trust that everything after it will be non-opt arguments.