Don't start tray menu if it's running already
Saves PID of the last command (yad) to a pidfile in the runtime dir and
checks against it each time tray_menu is called. We *could* also check
the existence of the traymenu fifo here, but since that can safely be
deleted while the tray menu is running, that could lead to false
positives. Instead when pidfile exists we the running PID.