Use basename for filename and destination
With file paths prefixed with $PWD, it prevents us from ever uploading
anything with a sensible path to the file. Doesn't work with ~, /, or
much else beyond paths relative to the current directory.

Instead we assign the basename to a new variable and use that, which
gives us the benefit of not trying to cp to $cache*foo/bar/baz.png. We
can also safely remove $PWD since the -E added in f5351a4 will ensure
the given file really does exist before going any further.