Alter main function to support differing outputs
Wraps the existing upload code in an if/else so we can bypass it when
the -c option is passed, instead copying directly to the clipboard.

`$image` is deceiving - it's actually the filename. So we pass that into
the clipboard helper function which will have xclip copy the file. For
that to work though, we need to specifically set the mime type.

May or may not work on Wayland, currently untested.