Fix regression in copying link to clipboard
Forgot that to_clipboard is also used for the text links, so clearly
setting the mime type to image/png is a bad idea most of the time.

This updates to always set the mime type, but defaults to text/plain.

The wl-copy functionality is guesswork. Assuming it's the same as paste.