Fix quit menu item
At some point Yad's built-in "quit" special action stopped working,
presumably due to a recent-ish update. My pull request to make it work
while padded with a space still hasn't been accepted and now that it
doesn't even work without it, I'm not sure that would change anything.

Replaces the quit action with a simple kill command, assigning the tray
process' PID to a local var first before it's written to the pidfile.