Show forced environment status in dep checker
Updates status_check to show the actual detected environment as well as
adding a new line with the "active" environment set by --env.

For the latter to work, we also need to make sure the ENV var is defined
before setting `chk` since status_check is called within the arg switch,
so we duplicate the `[ -n "..." ] || ...` bit that would otherwise run
before triggering the actual screenshot functionality.

autoset_environment must be called at the start as it's needed for the
is_wayland function to return the "correct" environment.