Support use of NEXTSHOT_ENV environment variable
Replaces NC_ENV with the more specific NEXTSHOT_ENV and moves parsing
code into its own function, incorporating the previous autoset function.

With this change, the environment can now be set by either exporting
NEXTSHOT_ENV in ~/.*shrc, or running with `NEXTSHOT_ENV=foo nextshot` in
addition to the `--env` flag.

Accepted values remain the same, with the exception that a blank or
unset value will be interpreted as "auto" so that an error isn't
triggered when running Nextshot without NEXTSHOT_ENV being set.