Add -v, --verbose option to enable debug mode
Updates the curl upload command to support redirecting output to a file
when debug mode is enabled so that it can be shown with the "Upload
failed" error to help troubleshoot issues such as #54.

While we could simply send output to STDERR with `-o /dev/stderr`, the
output ends up looking a bit off because there's no `\n` after the
progress bar, making it difficult to read the error.

In future it would be nice to parse the XML returned from Nextcloud, but
displaying the raw response should suffice for now.