Handle errors in nc_share function
Directs output of the curl share request into share.json, outputs it
to stderr if debug mode is enabled, and checks for a 200 response before
aborting in case for some reason upload works but sharing fails (say for
instance the OCS API changes at some point).

Also sets respCode as a local var in both upload and share since that
wasn't done before. Not sure why, but it is now.