Always output upload and share URLs to terminal
Updates nc_share to always output the share link, regardless of whether
notify_send is installed or not. It may be useful to re-copy the link
manually some point after the initial copy.

Also adds the gallery URL when upload completes, which also has spaces
replaced with %20 so that terminals that make links clickable don't see
it as two separate words and only link on the first part.