Encode spaces in upload request
Replaces spaces in the filename with `%20` to avoid issues on certain
servers where the URL gets borked somewhere along the line.

Only applies the fix to the upload URL, as the path field in the share
request is essentially a form element so %20 breaks the request. Using
the plus sign in place of spaces in the filename would make sense, but
Nextcloud saves the file as-is, i.e. `some+file+with+spaces.png`.