Handle 204 response from Nextcloud
Fixes an issue where uploading a screenshot that already exists (either
by renaming or using the `--file` option), Nextshot would receive an
unexpected 204 and bail before trying to share the image.

The file does get uploaded but overwrites the image that was there
previously, so this should really have extra handling to prompt first.

When sharing a file that's already shared, the same link is returned
with a 200 OK response. This is useful but could also be dangerous.