Fix pasting URL from clipboard in Alacritty
Due to an upstream bug where the clipboard library used by Alacritty is
unable to handle content that's had a target set, attempting to paste a
URL that was copied to clipboard by Nextshot would result in an error
and no text being pasted. Previously this would trigger a hang instead.

This works around the issue by simply not setting the text/plain target.
If the user is copying an image rather than share URL then we still set
the target so that applications will know how to handle it. Alacritty
will still error in that case, but it's to be expected in a terminal.