Only call find once
Since a user could sit at the "Continue?" prompt for any amount of time,
there's a good chance the list of files will be different by the time it
comes to actually removing them, so this reworks it to be called once.

Xargs isn't entirely necessary here, but it will help by splitting into
multiple commands automatically if argument lists end up being too long
at any point. The du command also seems to report the wrong size when
the file list is piped to it directly, although I'm not certain why.