Force use of `.` as decimal when converting ints
If the locale is set to somewhere that uses a comma (for example) as its
decimal separator, printf fails with an "invalid number" error in the
`int2dec` function. Doesn't seem to affect `int2hex`, but this updates
it to use the same system behaviour anyway, just in case.

Fixes #64