Support setting default image format in config
Adds a new `format` option in the config which will be used as the
default file extension when suggesting filenames. That is in turn used
by imagemagick's `import` command to determine the image format.

For Wayland, grim has a `-t` arg for setting the filetype to either jpeg
or png, so we pass that in automatically based on the config value.