Handle mixed and uppercase image formats
Previously if you specified JPG in config or the CLI arg, it would've
been ignored as it doesn't match the lowercase regex. We could set a
shell option to enable mixed case matches, but we'd have to restore it
anyway or it'd interfere with other comparisons.

This adds a short helper for validating an image format, and updates
load_config to set format in lowercase regardless of whether it's being
defined in the config file or CLI args.