Add is_jpeg helper, similar to is_format
Since we now have the format helper containing the regex, it makes sense
to replace the two jpeg regexes with a similar helper function. By the
time it's used we've already set the format in stone, so no need to
pass it in as a function param as we can use the $format var instead.