Enable window selection via Slurp on Sway, as on X
Rather than spitting out a list of windows either in the CLI or via Yad
and requiring the user select one, we can now pipe geometries directly
to Slurp for display as clickable regions thanks to emersion/slurp#28.

This replaces the original manual selection prompts (#45) with exactly
that - saving geometries to a var then piping that list into Slurp which
will spit out the matching geometry. Currently custom selections are
also possible and not ideal, so we may need to work around that somehow.