Cleanup duplicate pretty checks
This renames the `make_url` helper to `make_share_url` to make way for a
new `make_url` function that takes care of inserting "index.php" based
on the `pretty_urls` option in config, and whether the path begins with
a slash. If it does and pretty_urls is off, we add index.php.