Fix extra slash in share url
Due to the combined if statement where pretty_urls was checked in line
with the first character of the path, all paths would hit the else part
for Nextcloud installs with pretty URLs enabled.

This instead checks the inverse, so that "relative" paths are spat out
without index.php first. The pretty_urls option will only then be
checked if it *does* start with '/'. When enabled, the path is simply
appended to the server URL. Finally index.php is added if it's disabled.