Cleanup and invert loop
Sorts things into functions and runs the loop on the cache dir instead
to minimise the number of files the script needs to check. Nextcloud's
screenshots dir is likely to keep growing while the cache stays small.

Also adds some extra debug info and supports setting the Nextcloud root
and/or screenshots directory with `NC_ROOT` and `NC_IMAGE_DIR`.