Workaround Yad ignoring icon before tray loads
If the tray hasn't loaded yet, Yad will skip the call to `set_icon` and
the Nextshot tray will just get the default green flask icon instead.

Polybar usually has the tray loaded in a few seconds, so to workaround
the issue we just spam the command every half-second for 4 seconds ish.

Fixes #74