Don't warn until commands are actually used
Prior to this commit, a system missing slop would warn regardless of
whether the mode uses slop to select a geometry or simply needs import.

With a couple new helpers we can *require* a package, or simply *prefer*
it. For example if Slop isn't available, we can fall back to imagick
which has its own built-in (albeit limited and buggy) selections.

The problem with the latter case is—when falling back to `import`—the
warnings aren't printed until the import command has exited. To fix
that issue we simply `sleep 1` to allow for output before `import`.