Fix empty filename in GUI mode
The CLI prompt when run interactively would always default to the
original filename if no value was given, but if the default value is
removed from the Yad prompt then the result is `mv foo/filename foo/`.

This always defaults to the old name rather than relying on an `else`,
and fixes a shellcheck bracket issue with the file extension check.